+++ CHANGELOG.md
@@ -1,3 +1,10 @@
 # Changelog
 
-All notable changes to this project will be documented in this file.
+## [1.0.0] - 2025-01-27
+
+### Added
+- Event bus system for Unity 6
+- Three dispatchers: Unity, ThreadPool, Immediate
+- Custom dispatcher support
+- Async/sync event sending
+- Static handler registration